+++ astarte/device/__init__.py
@@ -41,4 +41,6 @@ from .exceptions import (
     InterfaceFileDecodeError,
     InterfaceNotFoundError,
     JWTGenerationError,
+    DeviceConnectingError,
+    DeviceDisconnectedError,
 )