+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.12.1] - Unreleased
+
+### Added
+- Adding or removing interfaces from a device while the device is connected.
+  If an interface is added or removed the new device introspection is immediately sent to Astarte.
+
 ### Fixed
 - False values on boolean endpoints for server owned interfaces are correctly processed.